[Refactor] Move Serde out of group link
