+++ be/src/serde/CMakeLists.txt
@@ -17,28 +17,14 @@ set(LIBRARY_OUTPUT_PATH "${BUILD_DIR}/src/serde")
 # where to put generated binaries
 set(EXECUTABLE_OUTPUT_PATH "${BUILD_DIR}/src/serde")
 
-ADD_BE_LIB(SerdeCore
-        column_array_serde.cpp
-        )
-
-target_link_libraries(SerdeCore PUBLIC
-        ColumnCore
-        TypesCore
-        Common
-        Base
-        Gutil
-        StarRocksGen
-        streamvbyte
-        )
-
 ADD_BE_LIB(Serde
+        column_array_serde.cpp
         encode_context.cpp
         compress_strategy.cpp
         protobuf_serde.cpp
         )
 
 target_link_libraries(Serde PUBLIC
-        SerdeCore
         RuntimeCore
         ChunkCore
         ColumnCore
@@ -47,4 +33,5 @@ target_link_libraries(Serde PUBLIC
         Base
         Gutil
         StarRocksGen
+        streamvbyte
         )
